Feat: 이미지,태그 dto에 인덱스 생성 생성 외

+++ src/main/java/spharos/nu/goods/domain/goods/service/GoodsService.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 import java.util.stream.IntStream;
 
 import lombok.extern.slf4j.Slf4j;
+
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.kafka.core.KafkaTemplate;
@@ -16,6 +17,8 @@ import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import spharos.nu.goods.domain.bid.service.BidService;
 import spharos.nu.goods.domain.goods.dto.event.CloseEventDto;
+import spharos.nu.goods.domain.goods.dto.request.ImageDto;
+import spharos.nu.goods.domain.goods.dto.request.TagDto;
 import spharos.nu.goods.domain.goods.dto.response.GoodsAllListDto;
 import spharos.nu.goods.domain.goods.dto.request.GoodsCreateDto;
 import spharos.nu.goods.domain.goods.dto.response.GoodsDetailDto;
@@ -28,6 +31,8 @@ import spharos.nu.goods.domain.goods.entity.Tag;
 import spharos.nu.goods.domain.goods.repository.GoodsRepository;
 import spharos.nu.goods.domain.goods.repository.ImageRepository;
 import spharos.nu.goods.domain.goods.repository.TagRepository;
+import spharos.nu.goods.global.exception.CustomException;
+import spharos.nu.goods.global.exception.errorcode.ErrorCode;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -42,7 +47,7 @@ public class GoodsService {
 	private final KafkaTemplate<String, OpenEventDto> openEventDtoKafkaTemplate;
 
 	public GoodsAllListDto goodsAllRead(Long categoryPk, boolean isTradingOnly, Pageable pageable) {
-		Page<GoodsCodeDto> goodsPage = goodsRepository.findAllGoods(categoryPk,isTradingOnly,pageable);
+		Page<GoodsCodeDto> goodsPage = goodsRepository.findAllGoods(categoryPk, isTradingOnly, pageable);
 		return GoodsAllListDto.builder()
 			.maxPage(goodsPage.getTotalPages())
 			.nowPage(goodsPage.getNumber())
@@ -78,17 +83,14 @@ public class GoodsService {
 				.goodsCode(goodsCode)
 				.build())
 		);
-
 		//이미지 저장
-		IntStream.range(0, goodsCreateDto.getImageUrls().size())
-			.forEach(index -> {
-				String imageUrl = goodsCreateDto.getImageUrls().get(index);
-				imageRepository.save(Image.builder()
-					.url(imageUrl)
-					.goodsCode(goodsCode)
-					.index(index)
-					.build());
-			});
+		goodsCreateDto.getImages().forEach((image) ->
+			imageRepository.save(Image.builder()
+				.url(image.getUrl())
+				.index(image.getId())
+				.goodsCode(goodsCode)
+				.build())
+		);
 
 		// 경매 시작시간에 스케줄링 걸기
 		scheduleOpenGoods(savedGoods);
@@ -112,8 +114,10 @@ public class GoodsService {
 	public GoodsDetailDto getGoodsDetail(String goodsCode) {
 		Goods goods = goodsRepository.findOneByGoodsCode(goodsCode).orElseThrow();
 
-		List<String> tags = tagRepository.findAllByGoodsCode(goodsCode).stream().map(Tag::getName).toList();
-		List<String> imageUrls = imageRepository.findAllByGoodsCode(goodsCode).stream().map(Image::getUrl).toList();
+		List<TagDto> tags = tagRepository.findAllByGoodsCode(goodsCode).stream()
+			.map((tag) -> TagDto.builder().id(tag.getId()).name(tag.getName()).build()).toList();
+		List<ImageDto> imageUrls = imageRepository.findAllByGoodsCode(goodsCode).stream()
+			.map((image) -> ImageDto.builder().id(image.getIndex()).url(image.getUrl()).build()).toList();
 
 		return GoodsDetailDto.builder()
 			.tradingStatus(goods.getTradingStatus())
@@ -144,9 +148,16 @@ public class GoodsService {
 
 	@Transactional
 	public Void goodsDelete(String goodsCode) {
+		/*상태가 0이 아닌 경우 삭제 불가능*/
+		Goods goods = goodsRepository.findOneByGoodsCode(goodsCode).orElseThrow();
+		if (goods.getTradingStatus() != 0) {
+			throw new CustomException(ErrorCode.FORBIDDEN_DELETE);
+		}
+
 		goodsRepository.deleteByGoodsCode(goodsCode);
 		tagRepository.deleteAllByGoodsCode(goodsCode);
 		imageRepository.deleteAllByGoodsCode(goodsCode);
+
 		return null;
 	}
 
@@ -165,8 +176,8 @@ public class GoodsService {
 			.sellerUuid(uuid)
 			.goodsCode(goods.getGoodsCode())
 			.closedAt(goods.getClosedAt())
-			.tradingStatus((byte)4) //거래취소로 변경
-			.isDelete(true) //삭제여부 true 로 변경
+			.tradingStatus(goods.getTradingStatus())
+			.isDisable(true) //숨김 여부 true 로 변경
 			.build();
 
 		goodsRepository.save(updatedGoods);
@@ -175,12 +186,12 @@ public class GoodsService {
 
 	// 경매 시작시간에 스케줄링
 	private void scheduleOpenGoods(Goods goods) {
-		taskScheduler.schedule(() -> OpenGoods(goods),Timestamp.valueOf(goods.getOpenedAt()));
+		taskScheduler.schedule(() -> OpenGoods(goods), Timestamp.valueOf(goods.getOpenedAt()));
 	}
 
 	@Transactional
 	public void OpenGoods(Goods goods) {
-		log.info("(상품 코드: {}) 경매 시작 ",goods.getGoodsCode());
+		log.info("(상품 코드: {}) 경매 시작 ", goods.getGoodsCode());
 
 		OpenEventDto openEventDto = OpenEventDto.builder()
 			.goodsCode(goods.getGoodsCode())
